fix: Correct typos in workspace.js descriptions
The commit fixes a typo in the descriptions of the Swagger responses in the workspace.js file. The word "toatl" has been corrected to "total" in all three responses.

The commit also removes unnecessary code in the getTotalBookingByManager function in the index.js file. The code related to fetching total workspaces has been removed.

Lastly, the commit updates the get.service.js file in the booking service. It adds a check to ensure that the manager exists before proceeding with the query. It also updates the where clause to use the manager_id from the manager object instead of the tokenUser.user_id.