refactor: Add getRevenueIn8DaysController
Added a new controller function, getRevenueIn8DaysController, to retrieve the revenue for the past 8 days. This controller handles the request to get the total prices of all bookings within the specified building. It is accessible to admin and manager roles. The function validates the building ID and returns the revenue if successful. Otherwise, it returns an error message.

Refactored files:
- src/controllers/booking/get.controller.js
- src/routes/v1/booking.js
- src/services/booking/get.service.js
- src/tests/index.js